Implementing graph algorithms with LLMs
Welcome to the first assignment of Course 1 in the Generative AI for Software Development Specialization! In this assignment you will work alongside an LLM to take on four different graph-based problems including finding the shortest path through a graph and solving different variations of the classic Travelling Salesman Problem in which you must find the shortest tour through every vertex in the graph. This is a great opportunity to practice using the LLM skills you've been learning! You can use the LLM of your choosing but GPT-4o is available here, in the ungraded lab that accompanies this assignment.

Here's a quick summary of the contents of this notebook:

Section 1: Introduction, setup, and the Graph class this assignment is based on
Section 2: The Graph_Advanced class that you will edit for your submission
Section 3: Description of the four exercises you'll implement inside of Graph_Advanced
Section 4: A playground to test solutions
Section 5: Unit tests of your Graph_Advanced class

Here's a few important points on how this notebook works:

1. Only Graph_Advanced is Graded: You can add new cells to experiment but these will be ignored by the grader. If you want something graded, include it in the provided cell that contains the Graph_Advanced class.
2. Some Cells are Frozen: Some cells are frozen, e.g. the Graph class, to avoid mistakenly editing their code
3. Avoid importing new libraries: Avoid importing new libraries beyond those you'll find in section 1.1, immediately below. In particular joblib or any multiprocessing or parallel computing approach will crash the grader. If you absolutely need additional libraries, import them in the same cell as the Graph_Advanced class.
4. Save before submitting: Do this to ensure you are graded on your most recent work.